refactoring system tests of Controllers
